feat(HolidayPicker): move picker to modal
